+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#1820** : Stroom now records `UploadedBy` meta attribute when manually uploading data via the UI.
+
+* Issue **#190** : Statistics are now recorded when manually uploading data.
+
 * Change stroom dependencies to use maven central instead of bintray.
 
 * Issue **#2207** : Fixed dashboard column rename issue where column name was not updated visually after rename.